+++ src/shaders/attention_aggregate_value.wgsl
@@ -7,9 +7,9 @@ struct Params {
   K_max: u32,
 };
 
-@binding(0) @group(0) var<storage, read_write> output : array<f32>; //L * dim
-@binding(1) @group(0) var<storage, read> V : array<f32>; //L * dim
-@binding(2) @group(0) var<storage, read_write> slate : array<f32>; //L * L
+@binding(0) @group(0) var<storage, read_write> output : array<f16>; //L * dim
+@binding(1) @group(0) var<storage, read> V : array<f16>; //L * dim
+@binding(2) @group(0) var<storage, read_write> slate : array<f16>; //L * L
 @binding(3) @group(0) var<uniform> params : Params;
 
 @compute @workgroup_size(1, 32)
