+++ src/shaders/silu_inplace.wgsl
@@ -4,7 +4,7 @@ struct Params {
   L: u32,
 };
 
-@binding(0) @group(0) var<storage, read_write> x : array<f32>;
+@binding(0) @group(0) var<storage, read_write> x : array<f16>;
 @binding(2) @group(0) var<uniform> params : Params;
 
 @compute @workgroup_size(1)
